Fix critical bug in number literal parsing that made it impossible to parse floating point literals.

+++ MobTalkerScript/src/mobtalkerscript/mts/v1/parser/MtsToMislCompiler.java
@@ -269,7 +269,7 @@ public class MtsToMislCompiler extends AbstractMtsToMislCompiler
     @Override
     public Void visitNumberLiteral( NumberLiteralContext ctx )
     {
-        int literal = Integer.parseInt( ctx.Literal.getText() );
+        double literal = Double.parseDouble( ctx.Literal.getText() );
         addInstr( new InstrPush( literal ) );
         
         return null;
